BeatAnimesBot/requirements.txt
pip
setuptools 
wheel
aiohttp
pyrogram>=2.0
TgCrypto
python-dotenv
feedparser
pillow
requests
bs4
uvloop; sys_platform != 'win32'
motor
dnspython
opencv-python-headless
aiofiles
techzapi